+++ src/utils.cpp
@@ -32,7 +32,7 @@ void printMat(cv::Mat m, std::string msg)
     {
         for (int j = 0; j < m.cols; j++)
         {
-            std::cout << m.at<double>(i, j) << "\t";
+            std::cout << m.at<float>(i, j) << "\t";
         }
         std::cout << std::endl;
     }
@@ -68,26 +68,26 @@ void printPoint2f(cv::Point2f m, std::string msg)
 cv::Mat getDistortionMatrix()
 {
     // Create zero distortion
-    cv::Mat distCoeffs(4, 1, cv::DataType<double>::type);
-    distCoeffs.at<double>(0) = 0;
-    distCoeffs.at<double>(1) = 0;
-    distCoeffs.at<double>(2) = 0;
-    distCoeffs.at<double>(3) = 0;
+    cv::Mat distCoeffs(4, 1, cv::DataType<float>::type);
+    distCoeffs.at<float>(0) = 0;
+    distCoeffs.at<float>(1) = 0;
+    distCoeffs.at<float>(2) = 0;
+    distCoeffs.at<float>(3) = 0;
     return distCoeffs;
 }
 
-cv::Mat getCameraMatrix(double fx, double fy, double cx, double cy)
+cv::Mat getCameraMatrix(float fx, float fy, float cx, float cy)
 {
-    cv::Mat cam(3, 3, cv::DataType<double>::type);
-    cam.at<double>(0, 0) = fx;
-    cam.at<double>(0, 1) = 0;
-    cam.at<double>(0, 2) = cx;
-    cam.at<double>(1, 0) = 0;
-    cam.at<double>(1, 1) = fy;
-    cam.at<double>(1, 2) = cy;
-    cam.at<double>(2, 0) = 0;
-    cam.at<double>(2, 1) = 0;
-    cam.at<double>(2, 2) = 1;
+    cv::Mat cam(3, 3, cv::DataType<float>::type);
+    cam.at<float>(0, 0) = fx;
+    cam.at<float>(0, 1) = 0;
+    cam.at<float>(0, 2) = cx;
+    cam.at<float>(1, 0) = 0;
+    cam.at<float>(1, 1) = fy;
+    cam.at<float>(1, 2) = cy;
+    cam.at<float>(2, 0) = 0;
+    cam.at<float>(2, 1) = 0;
+    cam.at<float>(2, 2) = 1;
     return cam;
 }
 
@@ -113,34 +113,34 @@ cv::Vec3f convertToVec3f(cv::Point3f &p)
 
 cv::Mat convertToMat(matd_t *m)
 {
-    cv::Mat mat(m->nrows, m->ncols, cv::DataType<double>::type);
+    cv::Mat mat(m->nrows, m->ncols, cv::DataType<float>::type);
     for (int i = 0; i < m->nrows; i++)
         for (int j = 0; j < m->ncols; j++)
         {
-            mat.at<double>(i, j) = matd_get(m, i, j);
+            mat.at<float>(i, j) = matd_get(m, i, j);
         }
     return mat;
 }
 
 cv::Mat convertToMat(std::vector<cv::Point3f> &v)
 {
-    cv::Mat m(3, v.size(), cv::DataType<double>::type);
+    cv::Mat m(3, v.size(), cv::DataType<float>::type);
     for (int i = 0; i < v.size(); i++)
     {
-        m.at<double>(0, i) = v[i].x;
-        m.at<double>(1, i) = v[i].y;
-        m.at<double>(2, i) = v[i].z;
+        m.at<float>(0, i) = v[i].x;
+        m.at<float>(1, i) = v[i].y;
+        m.at<float>(2, i) = v[i].z;
     }
     return m;
 }
 
 cv::Mat convertToMat(std::vector<cv::Point2f> &v)
 {
-    cv::Mat m(2, v.size(), cv::DataType<double>::type);
+    cv::Mat m(2, v.size(), cv::DataType<float>::type);
     for (int i = 0; i < v.size(); i++)
     {
-        m.at<double>(0, i) = v[i].x;
-        m.at<double>(1, i) = v[i].y;
+        m.at<float>(0, i) = v[i].x;
+        m.at<float>(1, i) = v[i].y;
     }
     return m;
 }
@@ -154,7 +154,7 @@ cv::Vec3f convertToVec3f(matd_t *m)
 }
 
 // defines a cube with side length = size * 2
-std::vector<cv::Point3f> defineCubeWithPoints(double size)
+std::vector<cv::Point3f> defineCubeWithPoints(float size)
 {
     std::vector<cv::Point3f> ret;
     ret.push_back(cv::Point3f(0, 0, 0));           // front bottom left
@@ -166,16 +166,16 @@ std::vector<cv::Point3f> defineCubeWithPoints(double size)
     ret.push_back(cv::Point3f(size, size, -size)); // back top right
     ret.push_back(cv::Point3f(0, size, -size));    // back top left
     // translate so that center is at the origin
-    cv::Mat t(3, 1, cv::DataType<double>::type);
-    t.at<double>(0, 0) = -size / 2;
-    t.at<double>(1, 0) = -size / 2;
-    t.at<double>(2, 0) = 0;
+    cv::Mat t(3, 1, cv::DataType<float>::type);
+    t.at<float>(0, 0) = -size / 2;
+    t.at<float>(1, 0) = -size / 2;
+    t.at<float>(2, 0) = 0;
     translatePoints(ret, t);
     return ret;
 }
 
 // TODO: change values
-std::vector<cv::Vec3f> defineCubeWithVectors(double side_length)
+std::vector<cv::Vec3f> defineCubeWithVectors(float side_length)
 {
     std::vector<cv::Vec3f> ret;
 
@@ -234,15 +234,15 @@ void rotatePoints(std::vector<cv::Point3f> &cube, cv::Mat R)
 {
     for (auto &point : cube)
     {
-        double tempx = point.x;
-        double tempy = point.y;
-        double tempz = point.z;
-        point.x = R.at<double>(0, 0) * tempx + R.at<double>(0, 1) * tempy +
-                  R.at<double>(0, 2) * tempz;
-        point.y = R.at<double>(1, 0) * tempx + R.at<double>(1, 1) * tempy +
-                  R.at<double>(1, 2) * tempz;
-        point.z = R.at<double>(2, 0) * tempx + R.at<double>(2, 1) * tempy +
-                  R.at<double>(2, 2) * tempz;
+        float tempx = point.x;
+        float tempy = point.y;
+        float tempz = point.z;
+        point.x = R.at<float>(0, 0) * tempx + R.at<float>(0, 1) * tempy +
+                  R.at<float>(0, 2) * tempz;
+        point.y = R.at<float>(1, 0) * tempx + R.at<float>(1, 1) * tempy +
+                  R.at<float>(1, 2) * tempz;
+        point.z = R.at<float>(2, 0) * tempx + R.at<float>(2, 1) * tempy +
+                  R.at<float>(2, 2) * tempz;
     }
 }
 
@@ -251,9 +251,9 @@ void translatePoints(std::vector<cv::Point3f> &cube, cv::Mat t)
 {
     for (auto &point : cube)
     {
-        point.x += t.at<double>(0, 0);
-        point.y += t.at<double>(1, 0);
-        point.z += t.at<double>(2, 0);
+        point.x += t.at<float>(0, 0);
+        point.y += t.at<float>(1, 0);
+        point.z += t.at<float>(2, 0);
     }
 }
 
@@ -265,7 +265,7 @@ void drawCube(std::vector<cv::Point3f> &cube, cv::Mat &frame,
     // Define a 3D transformation matrix that transforms coordinates from the
     // camera's coordinate system to the apriltag's coordinate system
     // Convert rotation and translation to 4x4 transformation matrix
-    cv::Mat transformationMatrix = cv::Mat::eye(4, 4, cv::DataType<double>::type);
+    cv::Mat transformationMatrix = cv::Mat::eye(4, 4, cv::DataType<float>::type);
     cv::Mat submatrix = transformationMatrix(cv::Rect(0, 0, 3, 3));
     rotationMatrix.copyTo(submatrix);
     translationMatrix.copyTo(transformationMatrix(cv::Rect(3, 0, 1, 3)));
@@ -309,7 +309,7 @@ void drawCube(std::vector<cv::Point3f> &cube, cv::Mat &frame,
              1); // front top left to back top left
 }
 
-std::vector<cv::Point3f> defineAxesWithPoints(double size)
+std::vector<cv::Point3f> defineAxesWithPoints(float size)
 {
     std::vector<cv::Point3f> pts;
     pts.push_back(cv::Point3f(0, 0, 0));    // origin
@@ -327,7 +327,7 @@ void drawAxes(std::vector<cv::Point3f> &axes, cv::Mat &frame,
     // Define a 3D transformation matrix that transforms coordinates from the
     // camera's coordinate system to the apriltag's coordinate system
     // Convert rotation and translation to 4x4 transformation matrix
-    cv::Mat transformationMatrix = cv::Mat::eye(4, 4, cv::DataType<double>::type);
+    cv::Mat transformationMatrix = cv::Mat::eye(4, 4, cv::DataType<float>::type);
     cv::Mat submatrix = transformationMatrix(cv::Rect(0, 0, 3, 3));
     rotationMatrix.copyTo(submatrix);
     translationMatrix.copyTo(transformationMatrix(cv::Rect(3, 0, 1, 3)));
@@ -355,19 +355,19 @@ void drawAxes(std::vector<cv::Point3f> &axes, cv::Mat &frame,
 
 std::array<float, 3> getRPY(cv::Mat &R)
 {
-    float sy = sqrt(R.at<double>(0, 0) * R.at<double>(0, 0) + R.at<double>(1, 0) * R.at<double>(1, 0));
+    float sy = sqrt(R.at<float>(0, 0) * R.at<float>(0, 0) + R.at<float>(1, 0) * R.at<float>(1, 0));
     bool singular = sy < 1e-6; // If
     float x, y, z;
     if (!singular)
     {
-        x = atan2(R.at<double>(2, 1), R.at<double>(2, 2));
-        y = atan2(-R.at<double>(2, 0), sy);
-        z = atan2(R.at<double>(1, 0), R.at<double>(0, 0));
+        x = atan2(R.at<float>(2, 1), R.at<float>(2, 2));
+        y = atan2(-R.at<float>(2, 0), sy);
+        z = atan2(R.at<float>(1, 0), R.at<float>(0, 0));
     }
     else
     {
-        x = atan2(-R.at<double>(1, 2), R.at<double>(1, 1));
-        y = atan2(-R.at<double>(2, 0), sy);
+        x = atan2(-R.at<float>(1, 2), R.at<float>(1, 1));
+        y = atan2(-R.at<float>(2, 0), sy);
         z = 0;
     }
     std::array<float, 3> ret;
@@ -380,15 +380,15 @@ std::array<float, 3> getRPY(cv::Mat &R)
 cv::Mat fromRPY(float roll, float pitch, float yaw)
 {
     // Calculate rotation about x axis
-    cv::Mat R_x = (cv::Mat_<double>(3, 3) << 1, 0, 0,
+    cv::Mat R_x = (cv::Mat_<float>(3, 3) << 1, 0, 0,
                    0, cos(roll), -sin(roll),
                    0, sin(roll), cos(roll));
     // Calculate rotation about y axis
-    cv::Mat R_y = (cv::Mat_<double>(3, 3) << cos(pitch), 0, sin(pitch),
+    cv::Mat R_y = (cv::Mat_<float>(3, 3) << cos(pitch), 0, sin(pitch),
                    0, 1, 0,
                    -sin(pitch), 0, cos(pitch));
     // Calculate rotation about z axis
-    cv::Mat R_z = (cv::Mat_<double>(3, 3) << cos(yaw), -sin(yaw), 0,
+    cv::Mat R_z = (cv::Mat_<float>(3, 3) << cos(yaw), -sin(yaw), 0,
                    sin(yaw), cos(yaw), 0,
                    0, 0, 1);
     // Combined rotation matrix
@@ -412,28 +412,28 @@ Transform constructTransform(cv::Mat R, cv::Mat t)
     return tf;
 }
 
-cv::Mat getRotationMatrix(double roll, double pitch, double yaw)
+cv::Mat getRotationMatrix(float roll, float pitch, float yaw)
 {
     cv::Mat R = cv::Mat::zeros(3, 3, CV_64F);
 
-    double c_roll = std::cos(roll);
-    double s_roll = std::sin(roll);
-    double c_pitch = std::cos(pitch);
-    double s_pitch = std::sin(pitch);
-    double c_yaw = std::cos(yaw);
-    double s_yaw = std::sin(yaw);
+    float c_roll = std::cos(roll);
+    float s_roll = std::sin(roll);
+    float c_pitch = std::cos(pitch);
+    float s_pitch = std::sin(pitch);
+    float c_yaw = std::cos(yaw);
+    float s_yaw = std::sin(yaw);
 
-    R.at<double>(0, 0) = c_yaw * c_pitch;
-    R.at<double>(0, 1) = c_yaw * s_pitch * s_roll - s_yaw * c_roll;
-    R.at<double>(0, 2) = c_yaw * s_pitch * c_roll + s_yaw * s_roll;
+    R.at<float>(0, 0) = c_yaw * c_pitch;
+    R.at<float>(0, 1) = c_yaw * s_pitch * s_roll - s_yaw * c_roll;
+    R.at<float>(0, 2) = c_yaw * s_pitch * c_roll + s_yaw * s_roll;
 
-    R.at<double>(1, 0) = s_yaw * c_pitch;
-    R.at<double>(1, 1) = s_yaw * s_pitch * s_roll + c_yaw * c_roll;
-    R.at<double>(1, 2) = s_yaw * s_pitch * c_roll - c_yaw * s_roll;
+    R.at<float>(1, 0) = s_yaw * c_pitch;
+    R.at<float>(1, 1) = s_yaw * s_pitch * s_roll + c_yaw * c_roll;
+    R.at<float>(1, 2) = s_yaw * s_pitch * c_roll - c_yaw * s_roll;
 
-    R.at<double>(2, 0) = -s_pitch;
-    R.at<double>(2, 1) = c_pitch * s_roll;
-    R.at<double>(2, 2) = c_pitch * c_roll;
+    R.at<float>(2, 0) = -s_pitch;
+    R.at<float>(2, 1) = c_pitch * s_roll;
+    R.at<float>(2, 2) = c_pitch * c_roll;
 
     return R;
 }
@@ -441,24 +441,24 @@ cv::Mat getRotationMatrix(double roll, double pitch, double yaw)
 cv::Mat getRotationFromTransform(Transform &tf)
 {
     cv::Mat R = cv::Mat::zeros(3, 3, CV_64F);
-    R.at<double>(0, 0) = tf.T.at<double>(0, 0);
-    R.at<double>(0, 1) = tf.T.at<double>(0, 1);
-    R.at<double>(0, 2) = tf.T.at<double>(0, 2);
-    R.at<double>(1, 0) = tf.T.at<double>(1, 0);
-    R.at<double>(1, 1) = tf.T.at<double>(1, 1);
-    R.at<double>(1, 2) = tf.T.at<double>(1, 2);
-    R.at<double>(2, 0) = tf.T.at<double>(2, 0);
-    R.at<double>(2, 1) = tf.T.at<double>(2, 1);
-    R.at<double>(2, 2) = tf.T.at<double>(2, 2);
+    R.at<float>(0, 0) = tf.T.at<float>(0, 0);
+    R.at<float>(0, 1) = tf.T.at<float>(0, 1);
+    R.at<float>(0, 2) = tf.T.at<float>(0, 2);
+    R.at<float>(1, 0) = tf.T.at<float>(1, 0);
+    R.at<float>(1, 1) = tf.T.at<float>(1, 1);
+    R.at<float>(1, 2) = tf.T.at<float>(1, 2);
+    R.at<float>(2, 0) = tf.T.at<float>(2, 0);
+    R.at<float>(2, 1) = tf.T.at<float>(2, 1);
+    R.at<float>(2, 2) = tf.T.at<float>(2, 2);
     return R;
 }
 
 cv::Mat getTranslationFromTransform(Transform &tf)
 {
     cv::Mat t = cv::Mat::zeros(3, 1, CV_64F);
-    t.at<double>(0, 0) = tf.T.at<double>(0, 3);
-    t.at<double>(1, 0) = tf.T.at<double>(1, 3);
-    t.at<double>(2, 0) = tf.T.at<double>(2, 3);
+    t.at<float>(0, 0) = tf.T.at<float>(0, 3);
+    t.at<float>(1, 0) = tf.T.at<float>(1, 3);
+    t.at<float>(2, 0) = tf.T.at<float>(2, 3);
     return t;
 }
 
@@ -474,32 +474,32 @@ void printTransform(Transform &tf, std::string msg)
     std::cout << "tf RPY: " << rpy[0] << ", " << rpy[1] << ", " << rpy[2] << std::endl;
 }
 
-cv::Mat getTranslationMatrix(double x, double y, double z)
+cv::Mat getTranslationMatrix(float x, float y, float z)
 {
     cv::Mat t = cv::Mat::zeros(3, 1, CV_64F);
-    t.at<double>(0, 0) = x;
-    t.at<double>(1, 0) = y;
-    t.at<double>(2, 0) = z;
+    t.at<float>(0, 0) = x;
+    t.at<float>(1, 0) = y;
+    t.at<float>(2, 0) = z;
     return t;
 }
 
-cv::Mat truncateVector(cv::Mat t, double fraction)
+cv::Mat truncateVector(cv::Mat t, float fraction)
 {
     cv::Mat shorter = t;
     for (int i = 0; i < shorter.rows; i++)
         for (int j = 0; j < shorter.cols; j++)
-            shorter.at<double>(i, j) = fraction * shorter.at<double>(i, j);
+            shorter.at<float>(i, j) = fraction * shorter.at<float>(i, j);
     return shorter;
 }
 
-double getAngularDifference(double th1, double th2)
+float getAngularDifference(float th1, float th2)
 {
     using namespace std;
 
     if (th1 == th2)
         return 0.0;
 
-    double diff = atan2((cos(th1) * sin(th2) - cos(th2) * sin(th1)),
+    float diff = atan2((cos(th1) * sin(th2) - cos(th2) * sin(th1)),
                         (sin(th1) * sin(th2) + cos(th1) * cos(th2)));
 
     while (diff > M_PI)
